+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [20.15.0]
+
+### Changed
+
+- bl64_msg_set_output: now accepts default value for output. Default will be ASCII for non-interactive shell, and ANSI for interactive
+
 ## [20.14.0]
 
 ### Changed